+++ apis_core/api_routers.py
@@ -27,7 +27,7 @@ from django_filters import rest_framework as filters
 from apis_core.apis_entities.models import TempEntityClass
 from .api_renderers import NetJsonRenderer
 from .apis_relations.models import Triple, Property
-from apis_core.helper_functions import caching
+from apis_core.utils import caching
 
 if "apis_highlighter" in getattr(settings, "INSTALLED_APPS"):
     from apis_highlighter.highlighter import highlight_text_new
