+++ apis_core/apis_metainfo/models.py
@@ -20,8 +20,8 @@ from django.utils.functional import cached_property
 from model_utils.managers import InheritanceManager
 
 # from django.contrib.contenttypes.fields import GenericRelation
-# from helper_functions.highlighter import highlight_text
-from apis_core.helper_functions import caching
+# from utils.highlighter import highlight_text
+from apis_core.utils import caching
 
 # from apis_core.apis_entities.serializers_generic import EntitySerializer
 # from apis_core.apis_labels.models import Label
@@ -33,7 +33,7 @@ if path_ac_settings:
     autocomp_settings = getattr(ac_settings, "autocomp_settings")
 else:
     from apis_core.default_settings.NER_settings import autocomp_settings
-# from apis_core.helper_functions import DateParser
+# from apis_core.utils import DateParser
 
 NEXT_PREV = getattr(settings, "APIS_NEXT_PREV", True)
