+++ apis_core/apis_relations/forms.py
@@ -6,7 +6,7 @@ from django import forms
 
 # from apis_core.apis_entities.models import Place
 from apis_core.apis_labels.models import Label
-from apis_core.helper_functions import DateParser
+from apis_core.utils import DateParser
 
 
 ##############################################
