+++ apis_core/urls.py
@@ -14,7 +14,7 @@ from apis_core.api_routers import views
 #     PlaceGeoJsonViewSet,
 # )
 # from apis_core.apis_vocabularies.api_views import UserViewSet
-from apis_core.helper_functions import caching
+from apis_core.utils import caching
 
 app_name = "apis_core"
